brikx_platform/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
supabase>=2.10.0
python-dotenv==1.0.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
lxml
requests